fix(Authentication): web config으로 bean 설정 전환 (1단계 끝)
